ебать накостылил в postPage fetchRemove id ругался пришлось костылить в postSlice и аллиуя работает спустя 3 часа блт

+++ CLIENT/src/components/Post/Post.tsx
@@ -31,9 +31,9 @@ const Post: React.FC<PostProps> = ({
 
   const dispatch = useAppDispatch()
   const navigate = useNavigate()
-
+  console.log('id :', id)
   const onClickRemove = () => {
-    if (window.confirm('Удалить пост?')) {
+    if (window.confirm('Удалить пост?') && id !== undefined) {
       dispatch(fetchRemovePost(id))
       navigate('/communism2.0')
     }
